backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
redis==5.0.4
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
polars==0.20.21
pyarrow==16.0.0
boto3==1.34.84
python-dotenv==1.0.1
celery==5.4.0
flower==2.0.1
aiofiles==23.2.1
